+++ demo/components/SplitModePreview.tsx
@@ -4,12 +4,13 @@ import transform from '@diplodoc/transform';
 import {useThemeValue} from '@gravity-ui/uikit';
 import type MarkdownIt from 'markdown-it';
 
-import {type MarkupString, colorClassName} from '../../src';
-import {debounce} from '../../src/lodash';
-import {YfmStaticView} from '../../src/view/components/YfmHtml';
-import {withLatex} from '../../src/view/hocs/withLatex';
-import {type MermaidConfig, withMermaid} from '../../src/view/hocs/withMermaid';
-import {withYfmHtmlBlock} from '../../src/view/hocs/withYfmHtml';
+import {type MarkupString, colorClassName} from 'src/index';
+import {debounce} from 'src/lodash';
+import {YfmStaticView} from 'src/view/components/YfmHtml';
+import {withLatex} from 'src/view/hocs/withLatex';
+import {type MermaidConfig, withMermaid} from 'src/view/hocs/withMermaid';
+import {withYfmHtmlBlock} from 'src/view/hocs/withYfmHtml';
+
 import {LATEX_RUNTIME, MERMAID_RUNTIME, YFM_HTML_BLOCK_RUNTIME} from '../defaults/md-plugins';
 import useYfmHtmlBlockStyles from '../hooks/useYfmHtmlBlockStyles';